change: Transaction details additions
Set domain and UI models for the details screen, also created mock TransactionClaimItems in the interactor and a transformer for the conversion of given items.
Screen was adjusted to support relevant info to be dislayed within a data shared section (for a list of attestations and their claims) and data signed section, using a WrapExpandableListItem in those cases to display an expandable card as shown in design.

Signed-off-by: Christos Kaitatzis <christos.kaitatzis@netcompany.com>