Feat: Adds support for Batch Issuance.
- Introduced `DocumentCredentialsInfo.kt` to model document credential
usage information.
- Updated `DocumentDetailsScreen.kt` and `DocumentDetailsViewModel.kt`
to display and manage an expandable section for document credential
information.
- Updated `DocumentsInteractor.kt` to display document credential counts
in the document list if the feature is enabled.
- Added new string resources for the document credentials info section.
- Updated Gradle dependencies for `eudiWalletCore` to `0.17.0-SNAPSHOT`
and modified the Maven repository URL for snapshots.
- Adjusted `WalletCorePresentationController.kt` to correctly handle
key unlock data and crypto objects for signing.
- Updated `WalletCoreDocumentsController.kt` to align with changes in
the core library, particularly regarding `getDefaultCreateDocumentSettings`
and user authentication for document issuance.
- Minor UI adjustments in `ListItem.kt` for trailing content.

- Replaced `DocumentMetaData` with `IssuerMetadata` in `DocumentHelper.kt`,
`DisplayExtensions.kt`, and `DocumentsExtensions.kt`.
- Updated `DocumentDetailsTransformer.kt` to use `IssuerMetadata`.
- Modified tests in `TestDocumentDetailsInteractor.kt`,
`TestDocumentsInteractor.kt`, and other test files to reflect the
metadata and all the other changes that the new Core version introduced (particularly the fact that IssuedDocument is no longer a data class.